requirements.txt
dj-database-url==2.3.0
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
django-filter==24.3
django-debug-toolbar==4.4.6
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dateutil
python-decouple==3.8
requests==2.32.3
whitenoise==6.8.2
pillow==11.0.0
